services/gitsync/requirements.txt
psycopg2-binary==2.9.3
SQLAlchemy==1.4.45
werkzeug==2.1.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
Deprecated==1.2.13
idna==3.4
pycparser==2.21
PyGithub==1.57
PyJWT==2.6.0
PyNaCl==1.5.0
requests==2.28.1
urllib3==1.26.13
wrapt==1.14.1
